Doppelgänger/requirements.txt
torch>=2.0.0  # Jetson wheel
onnxruntime-gpu>=1.15.0
sounddevice>=0.4.6
numpy>=1.21.0
scipy>=1.7.0
librosa>=0.9.0
pyaudio>=0.2.11
tensorrt>=8.6.0
onnx>=1.12.0
faiss-cpu>=1.7.0  # For RVC retrieval
matplotlib>=3.5.0  # For visualization if needed
tqdm>=4.64.0